Add reporting for Memory Limit Exceeded errors (status 12)
For `leetcode--show-submission-result` we're not showing status if
there's a Memory Limit Exceeded error (status 12).  This can happen,
for example, if a submission in Python hits the Python recursion depth
limit.  When it happens, nothing is displayed in the submission result
buffer at all.

In this change, we add some handling to deal with this error for the
submission results.