Add reporting for Output Limit Exceeded errors (status code 13)
The Leetcode system has a limit on how much output (i.e. to
stdout/stderr) a problem submission can produce.  If a submission does
produce too much output (e.g. due to print statement debugging left in
by accident), Leetcode returns an error with `status_code` 13.

Currently, the `leetcode--show-submission-result` function does not
detect errors with this status code, resulting in an empty
`leetcode-result-XXX` buffer and no indication of failure.  This
change adds detection of the output limit breach and notifies the user
in the results buffer that there was a problem.  Note that we do not
list the problem that was running at the time of the error, since
that's probably not relevant to the problem of outputting to stdout et
al.